actually BackendSIMT is a better place for ``genInitReductionTargets``
