+++ include/genn/genn/code_generator/backendBase.h
@@ -18,7 +18,6 @@
 #include "codeStream.h"
 #include "gennExport.h"
 #include "gennUtils.h"
-#include "varAccess.h"
 #include "variableMode.h"
 
 // Forward declarations
@@ -502,22 +501,6 @@ public:
 
 protected:
     //--------------------------------------------------------------------------
-    // ReductionTarget
-    //--------------------------------------------------------------------------
-    //! Simple struct to hold reduction targets
-    struct ReductionTarget
-    {
-        ReductionTarget(const std::string &n, const std::string &t, VarAccessMode a)
-        : name(n), type(t), access(a)
-        {
-        }
-
-        const std::string name;
-        const std::string type;
-        const VarAccessMode access;
-    };
-
-    //--------------------------------------------------------------------------
     // Protected API
     //--------------------------------------------------------------------------
     void addType(const std::string &type, size_t size, const std::string &lowestValue = "")
@@ -536,34 +519,6 @@ protected:
     void genSynapseIndexCalculation(CodeStream &os, const SynapseGroupMergedBase &sg, unsigned int batchSize) const;
 
     void genCustomUpdateIndexCalculation(CodeStream &os, const CustomUpdateGroupMerged &cu, unsigned int batchSize) const;
-
-    template<typename G>
-    std::vector<ReductionTarget> genInitReductionTargets(CodeStream &os, const G &cg) const
-    {
-        // Loop through variables
-        std::vector<ReductionTarget> reductionTargets;
-        const auto *cm = cg.getArchetype().getCustomUpdateModel();
-        for(const auto &v : cm->getVars()) {
-            // If variable is a reduction target, define variable initialised to correct initial value for reduction
-            // **NOTE** by not initialising this, compilers should emit a warning if user code doesn't set it to something
-            if(v.access & VarAccessModeAttribute::REDUCE) {
-                os << v.type << " lr" << v.name << " = " << getReductionInitialValue(*this, getVarAccessMode(v.access), v.type) << ";" << std::endl;
-                reductionTargets.emplace_back(v.name, v.type, getVarAccessMode(v.access));
-            }
-        }
-
-        // Loop through variable references
-        for(const auto &v : cm->getVarRefs()) {
-            // If variable reference is a reduction target, define variable initialised to correct initial value for reduction
-            // **NOTE** by not initialising this, compilers should emit a warning if user code doesn't set it to something
-            if(v.access & VarAccessModeAttribute::REDUCE) {
-                os << v.type << " lr" << v.name << " = " << getReductionInitialValue(*this, v.access, v.type) << ";" << std::endl;
-                reductionTargets.emplace_back(v.name, v.type, v.access);
-            }
-        }
-        return reductionTargets;
-    }
-
 
 private:
     //--------------------------------------------------------------------------
